Added Docker Support
This push includes the addition of various docker elements to create our container and run it. It is still in progress, do not attempt to change at the moment, it will break. Updates are still needed to run the front and backend at the same time using the unfinished docker-compose.yml.